+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [1.3.1] - 2022-07-19
+
+### Fixed
+
+- Try to decode content if empty
+- Trimmed content before inflate failed
+
 ## [1.3.0] - 2021-11-26
 
 ### Added